Correction in Rules permission
As per the above "Rights required for Service Bus operations" table, we need "Manage" claim to create or delete the rule. But we are able to create or delete rule with Listen privilege using 'SubscriptionClient.AddRule(RuleDescription)/SubscriptionClient.RemoveRule(string)' method. I guess the documentation should be corrected, so proposing this correction. Please review and apply.